backend/requirements.txt
# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0

# Database
supabase>=1.0.3
postgrest>=0.10.8
httpx>=0.24.0,<0.26.0

# Environment and utilities
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3

# Date handling
python-dateutil==2.8.2

# Logging and error handling
structlog==24.1.0

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.1.3
scikit-learn>=1.3.2
transformers>=4.36.0
python-multipart>=0.0.6
pytest>=7.4.3
selenium>=4.15.2
webdriver-manager>=4.0.1
numpy>=1.26.2
tqdm>=4.66.1
pydantic>=2.5.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
undetected-chromedriver>=3.5.5 